feat: refactor and modularize type extraction logic by introducing new parser utilities and restructuring component handling
