+++ packages/docs-extractor/src/handler/external-packages-handler.ts
